Added cache for previously know sensor values
Upon testing with Shelly BLU H&T there seems to be an issue that temperature and humidity values are missing every 6 hours.
As the characteristics cannot be set to null/undefined, currently the values are set to some absurd values to indicate a problem.

With this commit the last known sensor values are stored and used as characteristics values in case they are missing in the new paylod.
This doesn't apply for sporadic events such as the button press.